+++ .gitignore
@@ -1,7 +1,4 @@
-embedings_core/data/
 env/
-data/big_data/
-data/t-SNE_projected
-data/train_embeddings.npy
-data/TSNE_train_embeddings_clusters.csv
 data/
+.vscode/
+note.md
