--- clustering_core/include/t-sne/qdtsne.hpp
@@ -1,23 +1,0 @@
-#ifndef QDTSNE_QDTSNE_HPP
-#define QDTSNE_QDTSNE_HPP
-
-/**
- * @file qdtsne.hpp
- *
- * @brief Umbrella header for all **qdtsne** files.
- *
- * This header is effectively an alias for `tsne.hpp`.
- * It is provided just to allow developers to `#include "qdtsne/qdtsne.hpp"` without thinking too hard about the code-containing headers.
- * Of course, `#include qdtsne/tsne.hpp` also works but is less obvious.
- */
-
-#include "tsne.hpp"
-#include "utils.hpp"
-
-/**
- * @namespace qdtsne
- * @brief Quick and dirty t-SNE functions.
- */
-namespace qdtsne {}
-
-#endif
