+++ src/tests_core/output/test_result.csv
@@ -1,0 +1,3 @@
+cluster_id,distance,
+-1,0,
+-1,0,
