+++ src/visualizing/tempCodeRunnerFile.py
@@ -1,0 +1,2 @@
+fig = px.scatter(data, x='x', y='y', color='cluster', template='plotly_dark')
+# fig.show()
