--- visualizing/tempCodeRunnerFile.py
@@ -1,2 +1,0 @@
-fig = px.scatter(data, x='x', y='y', color='cluster', template='plotly_dark')
-# fig.show()
